Remove not_null magic modifier in favor of a composible one
This removes some magic which was also not very re-usable outside of the pre-built helpers. The only downside is that this may introduce some surprise, since typically, one would expect right-to-left execution by appearance. Also, the code "reads" right to left in a flow from source to target. Consistency would seem to dictate `value(not_null('source'), 'default')` - but that seems to require making the helpers themselves more complex.